==> Step1_Data_Pull_and_Clean.ipynb <==
### Imported Libraries

import os
import math
import pickle
import numpy as np
import pandas as pd
from os import listdir
from os.path import isfile, join

import pymatgen as mg
from pymatgen import MPRester
from pymatgen.io.cif import CifParser
from pymatgen.io import ase
from pymatgen.transformations.standard_transformations import OxidationStateDecorationTransformation
from pymatgen.transformations.standard_transformations import AutoOxiStateDecorationTransformation
from pymatgen.symmetry.groups import *
from pymatgen.symmetry.analyzer import SpacegroupAnalyzer

from ase import atoms
from ase.io import read, write

from pprint import pprint
from tqdm import notebook as tqdm
from tqdm.auto import tqdm as tqdm_pandas
tqdm_pandas.pandas()
# import swifter

### API KEY for Materials Project

__To run the notebook, a Materials Project API key is required!__ API keys can be generated [here](https://materialsproject.org/open).

mpr = MPRester(api_key='insert_key')

### 1a. Read in all Li-containing compounds from Materials Project
Operations come from mongodb. The syntax can be found [here](https://docs.mongodb.com/manual/reference/operator/query/).

Queryable parameters from Materials Project can be found [here](https://materialsproject.org/docs/api).

data = mpr.query(criteria={"elements": {"$in": ["Li"]}, "nelements": {"$gte": 2}}, 
               properties=["icsd_ids", "cif", "material_id", "pretty_formula", "spacegroup.symbol", "structure", "band_gap", "e_above_hull"])

The MPRester query returns a list of dictionaries, where each dictionary contains the properties of interest. Most of the data cleaning can be handled using the [Python Materials Genomics (pymatgen) library](https://pymatgen.org/). However, some operations are more conveniently accomplished via the [Atomic Simulation Environment (ASE)](https://wiki.fysik.dtu.dk/ase/index.html). We'll make a redundant list which contains the structures as [ase.atoms.Atoms](https://wiki.fysik.dtu.dk/ase/ase/atoms.html) type.  

AAA = ase.AseAtomsAdaptor
pymatgen_structures = []
ase_structures = []

for material in tqdm.tqdm(data):
    pymatgen_structures.append((material["structure"], material["icsd_ids"], material["material_id"], material["pretty_formula"], material["spacegroup.symbol"], material["band_gap"], material["e_above_hull"]))
    ase_structures.append(AAA.get_atoms(material["structure"])) 

### 1b. Read in all Li-containing compounds from ICSD
The ICSD Li-containing .cif files are contained in a seperate folder. Due to copyright, the cifs themselves are not shared in this data repository. The cifs have already been ordered using Pymatgen. Due to computational resources, about 10% of the cifs could not be ordered. 

path = os.path.join(os.getcwd(), 'Li_Cifs_folder')

# keep track of any errors here
errors = []

# all files in the directory are .cifs
for filename in tqdm.tqdm(os.listdir(path)):
    filepath = os.path.join(path, filename)
    try: 
        # the ICSD tag is being grabbed with filename[:-4]
        pymatgen_structures.append((CifParser(filepath).get_structures()[0], [int(filename[:-4])]))
        ase_structures.append(read(filepath))
    except:
        errors.append(filepath)

Remove files with errors after manually checking them. 

for item in errors[1:]:
    os.remove(item)

Check that the pymatgen_structures and ase_structures lists contain the same number of structures. 

len(pymatgen_structures)==len(ase_structures)

### 1c. Find all dupliciate ICSD IDs and Merge

Some of the Materials Project entries point to ICSD IDs. To avoid having duplicate data, we'll find these IDs and merge them. The latest duplicate will become the master because this ensures that ICSD data (experimental) is conserved. 

# list for IDs that will be deleted
deletion_list = []
# list for mapping the ICSD_IDs from Materials Project to those from the ICSD
mapping_IDs = []

# Loop over all the structures in pymatgen_structures
for idx, val in enumerate(tqdm.tqdm(pymatgen_structures, desc='Finding duplicate ICSD_IDs')):
    # Check all the ICSD IDs for each structures - there can and often are multiple listed IDs
    for item in val[1]:
        # Loop over all structures in pymatgen_structures
        for edx, val2 in enumerate(pymatgen_structures):
            # if the ICSD ID is in the list
            if item in val2[1]:
                # if the ICSD ID is a later entry then mark the earlier entry for deletion
                if edx > idx:
                    # mark the earlier entry for deletion
                    deletion_list.append(idx)
                    
                    # preserve the associated Materials Project values in the mapping_IDs list
                    # these vals are just the locations for ICSD_ID, MP_ID, spacegroup, bandgap, and Ehull
                    mapping_IDs.append((val[1], val[2], val[4], val[5], val[6]))


print('There are {} entries marked for deletion.'.format(len(deletion_list)))

### 1d. Mark some additional boundary cases for deletion

Structures that contain deuterium are not of interest. Find and mark for deletion. 

deuterium_containing = 0
for idx, val in enumerate(tqdm.tqdm(pymatgen_structures, desc='Check for deuterium')):
    for element in val[0].symbol_set:
        if element == 'D':
            deletion_list.append(idx)
            deuterium_containing+=1
            
print('There are {} entries that contain deuterium.'.format(deuterium_containing))

Structures with over 500 distinct lattice positions would be very difficult to work with. Mark for deletion. 

greater_than_500 = 0
for idx, structure in enumerate(tqdm.tqdm(ase_structures, desc='Checking for structures with over 500 sites')):
    if structure.get_global_number_of_atoms() > 500:
        greater_than_500 += 1
        deletion_list.append(idx)

print("There are {} structures with >500 atoms".format(greater_than_500))

Special case deletion: The structure $Li_7C_{120}$ (mp-1223102) throws errors. We'll just manually delete it. 

for idx, structure in enumerate(tqdm.tqdm(pymatgen_structures, desc='Checking for mp-1223102')):
    try:
        if structure[2] == 'mp-1223102':
            print("Structure mp-1223102 is at index {}.".format(idx))
            deletion_list.append(idx)
    except:
        pass

### 1e. Remove unwanted structures. Keep list of deleted structures for later. 

deleted_structures = [value for index, value in enumerate(pymatgen_structures) if index in deletion_list]
pymatgen_structures_sanitized = [value for index, value in enumerate(pymatgen_structures) if index not in deletion_list]
ase_structures_sanitized = [value for index, value in enumerate(ase_structures) if index not in deletion_list]

Check that the new lists are equivalent in length:

print("pymatgen_structures_sanitized and ase_structures_sanitized are equivalent length check:", len(pymatgen_structures_sanitized) == len(ase_structures_sanitized))
print("There are a total of {} structures.".format(len(pymatgen_structures_sanitized)))

### 1f. Reformat the sanitized lists into a composite dataframe (structures_df)

structures_df = pd.DataFrame(pymatgen_structures_sanitized)
structures_df.columns = ['structure','ICSD_ID', 'MP_ID', 'pretty_formula', 'spacegroup', 'bandgap', 'e_hull']
structures_df['ase_structure'] = ase_structures_sanitized
structures_df['composition'] = None
structures_df['replacement'] = None

# grab the output of the formula function and store it in the 'composition' column
structures_df.composition = structures_df.structure.map(lambda x: x.formula)
structures_df.ICSD_ID = structures_df.ICSD_ID.map(lambda x: int(x[0]) if (len(x)>0) else 0)

structures_df.head(20)

Label ICSD data with corresponding MP_IDS from the deleted compounds (saved in mapping_IDs).

# count the number of mapping events
count = 0

# for each line in the mapping_IDs list
for entry in tqdm.tqdm(mapping_IDs, desc='Label ICSD structures with the corresponding MP ID'):
    # for each value in the ICSD_IDs column
    for idx, value in enumerate(structures_df['ICSD_ID']):
        # if the value exists in the list of ICSD_IDs from the mapping_IDs
        if value in entry[0]:
            # if the structure doesn't already have a Materials Project ID
            if structures_df.loc[idx, 'MP_ID'] is None:
                structures_df.at[idx, 'MP_ID'] = entry[1]
                structures_df.at[idx, 'spacegroup'] = entry[2]
                structures_df.at[idx, 'bandgap'] = entry[3]
                structures_df.at[idx, 'e_hull'] = entry[4]
                count+=1

print("Materials project IDs added for {} structures.".format(count))
                

Add the deleted structures back in as a possible replacment column.

# count the number of replacment events
count = 0

# for each value in the ICSD_IDs column
for idx, ICSD_value in enumerate(tqdm.tqdm(structures_df['ICSD_ID'], desc='Add deleted structures into a replacement column')):
    # for each line in the deleted structures list
    for idx_del, deleted_structure in enumerate(deleted_structures):
        # if the ICSD value corresponds to a value in the deleted structures ICSD list
        if ICSD_value in deleted_structure[1]:
            structures_df.at[idx, 'replacement'] = deleted_structure[0]
            count+=1
            
print("{} deleted structures were added back into the 'replacement' column.".format(count))

structures_df.head(20)

### 1g. Check to see if the structures are ordered and oxidation state decorated

unordered_count = 0
unordered = []
no_oxidation_count = 0
no_oxidation_state = []

for idx in tqdm.tqdm(np.arange(0, len(structures_df), 1), desc='Check for oxidation states and ordering'):
    if not structures_df.loc[idx, "structure"].is_ordered:
        unordered_count += 1
        unordered.append(idx)
    try:
        for site in structures_df.loc[idx, "structure"]:
            _ = site.specie.oxi_state
    except Exception as e:
        no_oxidation_count +=1
        no_oxidation_state.append(idx)
        
print("{} structures are disordered.".format(unordered_count))
print("{} structures are not charge decorated.".format(no_oxidation_count))

### 1h. Apply charge decoration where possible. 

Pymatgen contains three general ways to charge decorate structures:
* Use the .add_oxdiation_state_by_guess method on a pymatgen structure. 
* Use the AutoOxiStateDecorationTransformation class.
* Create a dictionary of charges and apply based on atomic identity using the OxidationStateDecorationTransformation class.

# oxidation dictionary added to brute force oxidation states
oxidation_dictionary = {"H":1, "Li": 1, "Na":1, "K":1, "Rb": 1, "Cs":1, "Be":2, "Mg":2, "Ca":2, \
                        "Sr":2, "Ba":2, "Ra": 2, "B":3, "Al":3, "Ga":3, "In":3, "Tl":3, \
                        "C":4, "Si": 4, "Ge": 4, "Sn": 4, "Pb": 4, "N":-3, "P":5, "As":5, \
                        "Sb": 5, "Bi":5, "O":-2, "S":-2, "Se":-2, "Te":-2, "Po":-2, "F":-1, \
                       "Cl":-1, "Br":-1, "I":-1, "Sc":3, "Y":3, "Lu":3, "Ti":4, "Zr":4, "Hf":4, \
                       "V":5, "Nb":5, "Ta":5, "Cr":6, "Mo":4, "W":6, "Mn":7, "Tc":7, "Re":7, \
                       "Fe":3, "Ru":3, "Os":3, "Co": 3, "Rh":3, "Ir":3, "Cu":2, "Ag":1, "Au":3, \
                       "Zn":2, "Ni":2, "Cd":2, "Hg":2, "La":3, "Ce":3, "Pd":2, "Pm":3, "Ho":3, \
                        "Eu":3, "Np":3, "Pu":4, "Gd":3, "Sm":2, "Tb":3, "Tm":3, "Yb":3, "Ac":3, \
                       "Dy": 3, "Er":3, "Pr":3, "U":6, "Pt":2, "Nd":3, "Th":4, "Pa":5}

# two types of transformations taken from the pymatgen
oxidation_decorator = OxidationStateDecorationTransformation(oxidation_dictionary)
oxidation_auto_decorator = AutoOxiStateDecorationTransformation(distance_scale_factor=1)

def apply_charge_decoration(structure):
    """
    Function intended to be used with pandas.DataFrame.apply()

    For each structure passed into the function, it sequentially attempts three charge decoration strategies:
    (1) a manual charge decoration using the oxidation_dictionary
    (2) charge decoration using OxidationStateDecorationTransformation
    (3) charge decoration using the built-in add_oxidation_state_by_guess() method
    
    If any of these strategies result in a structure with nearly zero charge (<= 0.5), then the decoration is accepted.

    Parameters
    ----------
    structure : pymatgen.core.structure
        a pymatgen structure file

    Returns
    ------
    structure: pymatgen.core.structure
        either charge decorated or not (if the decorations failed)
    """
    
    # try the manual decoration strategy
    temp_structure = structure.copy()
    try:
        manually_transformed_structure = oxidation_decorator.apply_transformation(temp_structure)
        if abs(manually_transformed_structure.charge) < 0.5:
            return manually_transformed_structure
    except:
        pass
    
    # try Pymatgen's auto decorator
    temp_structure = structure.copy()
    try:
        auto_transformed_structure = oxidation_auto_decorator.apply_transformation(temp_structure)
        if abs(auto_transformed_structure.charge) < 0.5:
            return auto_transformed_structure
    except:
        pass
    
    # allow Pymatgen to guess the oxidation states
    temp_structure = structure.copy()
    try:
        structure.add_oxidation_state_by_guess()
        return structure
    except:
        pass 

    return structure

mask = [True if x in no_oxidation_state else False for x in structures_df.index]
structures_uncharged_df = structures_df[mask]
structures_df.loc[mask, 'structure'] = structures_uncharged_df['structure'].progress_apply(apply_charge_decoration)

unordered_count = 0
unordered = []
no_oxidation_count = 0
no_oxidation_state = []

for idx in tqdm.tqdm(np.arange(0, len(structures_df), 1), desc='Find the number of structures that remain disordered and/or charge undecorated.'):
    if not structures_df.loc[idx, "structure"].is_ordered:
        unordered_count += 1
        unordered.append(idx)
    try:
        for site in structures_df.loc[idx, "structure"]:
            _ = site.specie.oxi_state
    except Exception as e:
        no_oxidation_count +=1
        no_oxidation_state.append(idx)
        
print("{} structures are disordered.".format(unordered_count))
print("{} structures are not charge decorated.".format(no_oxidation_count))

### 1i. Save the dataframe using pickle

save_path = os.path.join(os.getcwd(), 'semi-supervised_supporting_files/structures_df_3p8_post_sanitize.pkl')
save_file = open(save_path, 'wb')
pickle.dump(structures_df, save_file)
save_file.close()

### 1j. Save deleted structures for future use

save_path_2 = os.path.join(os.getcwd(), 'semi-supervised_supporting_files/structures_deleted_post_sanitize.pkl')
save_file = open(save_path_2, 'wb')
pickle.dump(deleted_structures, save_file)
save_file.close()

### 1k. Test to make sure data loads well

save_path = os.path.join(os.getcwd(), 'semi-supervised_supporting_files/structures_df_3p8_post_sanitize.pkl')
open_file = open(save_path, 'rb')
structures_df = pickle.load(open_file)
open_file.close()

### 1l. Removal of structures that couldn't be charge decorated

unordered_count = 0
unordered = []
no_oxidation_count = 0
no_oxidation_state = []

for idx in tqdm.tqdm(np.arange(0, len(structures_df), 1), desc='Remove structures without charge decoration'):
    if not structures_df.loc[idx, "structure"].is_ordered:
        unordered_count += 1
        unordered.append(idx)
    try:
        for site in structures_df.loc[idx, "structure"]:
            _ = site.specie.oxi_state
    except Exception as e:
        no_oxidation_count +=1
        no_oxidation_state.append(idx)
        
print("{} structures are disordered.".format(unordered_count))
print("{} structures are not charge decorated.".format(no_oxidation_count))

for_removal = np.unique(unordered+no_oxidation_state)
structures_df = structures_df.drop(for_removal)
structures_df = structures_df.reset_index()

structures_df.head(-50)

### 1m. Routine to label ICSD structures with spacegroup

def spacegroup_labeler(structure):
    """
    Function intended to be used with pandas.DataFrame.apply()

    For each structure passed into the function, determine the space group.

    Parameters
    ----------
    structure : pymatgen.core.structure
        a pymatgen structure file

    Returns
    ------
    sg.get_space_group_symbol() : string
        the symbol for the spacegroup
    """
    
    try:
        s = mg.Structure.from_sites(structure)
        sg = SpacegroupAnalyzer(s)
        return sg.get_space_group_symbol()
    except:
        return None

missing_spacegroup_idx = structures_df[structures_df["spacegroup"].isna()].index.tolist()
print("Initial: {} rows are missing a spacegroup.".format(len(missing_spacegroup_idx)))

# make the dataframe to apply the spacegroup analyzer only to rows missing a spacegroup
mask = [True if x in missing_spacegroup_idx else False for x in structures_df.index]
structures_without_spacegroup_df = structures_df[mask]
structures_df.loc[mask, 'spacegroup'] = structures_without_spacegroup_df['structure'].progress_apply(spacegroup_labeler)


missing_spacegroup_idx = structures_df[structures_df["spacegroup"].isna()].index.tolist()
print("Post spacegroup labeling: {} rows are missing a spacegroup.".format(len(missing_spacegroup_idx)))

### 1n. Breakdown of spacegroups

spacegroup_counts = []
for i in np.arange(1, 231, 1):
    symbol = sg_symbol_from_int_number(i)
    print("{} - {}: {}".format(i, symbol, len(structures_df[structures_df["spacegroup"] == symbol])))
    spacegroup_counts.append(len(structures_df[structures_df["spacegroup"] == symbol]))

### 1o. Save a copy without ASE structures for 3.7 compatibility (comment out if not needed)

# savePath = os.path.join(os.getcwd(), 'structuresDF_3p8_postSanitize2_noASE.pkl')
# saveFile = open(savePath, 'wb')
# pickle.dump(newInstance.drop(['aseStructures'], axis=1), saveFile)
# saveFile.close()

### 1p. Create structure simplifications. Nine versions of the structure are used:

* __structure:__ The original structure. 
* __structure_A:__ Only the anions. All anions are modeled as S atoms. 
* __structure_AM:__ Only the anions and mobile atoms. Anions and mobil atoms are modeled as S and Li, respectively. 
* __structure_CAN:__ Only the cations, anions, and mobile atoms. Cations, anions, and mobile atoms are modeled as Al, S, and Li, respectively. 
* __structure_CAMN:__ All atoms are retained. Cations, anions, mobile atoms, and neutral atoms are modeled as Al, S, Li, and Mg, respectively. 
* __structure_A40:__ As with structure_A but the lattice volume is scaled to 40 cubic angstroms per anion. 
* __structure_AM40:__ As with structure_AM but the lattice volume is scaled to 40 cubic angstroms per anion. 
* __structure_CAN40:__ As with structure_CAN but the lattice volume is scaled to 40 cubic angstroms per anion. 
* __structure_CAMN40:__ As with structure_CAMN but the lattice volume is scaled to 40 cubic angstroms per anion. 

    

def structure_simplifications(structure_input, simplification_dict):
    """
    Function intended to be used with pandas.DataFrame.apply()

    The function takes a structure and then simplifies it based on the contents of simplification_dict.
    The simplified structure is returned. 


    Parameters
    ----------
    structure_input : pymatgen.core.structure
        a pymatgen structure file
        
    simplification_dict : dictionary
        A dictionary containing boolean values for the keys: 'C', 'A', 'M', 'N', '40'

    Returns
    ------
    structure: pymatgen.core.structure
        the simplified structure
    """
    
    # copy the structure in case modification fails
    structure = structure_input.copy()
    
    # create lists to keep track of the indices for the different atom types: cation, anion, mobile, neutral
    cation_list = []
    anion_list = []
    mobile_list = []
    neutral_list = []
    
    # create list to keep track of which atoms will be removed
    removal_list = []
    
    # integer to keep track of how to scale the lattice (for the representations that end in '40')
    scaling_counter = 0
    
    for idx, site in enumerate(structure):
        # grab the element name at the site
        element = site.species.elements[0].element.name
        # grab the charge at the site
        charge = site.specie.oxi_state
        
        # if the site is the mobile atom
        if element == 'Li':
            mobile_list.append(idx)
        else:
            # if the site holds a neutral atom
            if charge == 0:
                neutral_list.append(idx)
                scaling_counter+=1
                structure.replace(idx, mg.Specie("Mg", oxidation_state=charge))
            # if the site holds a cation
            elif charge>0:
                cation_list.append(idx)
                structure.replace(idx, mg.Specie("Al", oxidation_state=charge))
            # if the site holds an anion
            else:
                anion_list.append(idx)
                scaling_counter+=1
                structure.replace(idx, mg.Specie("S", oxidation_state=charge))
    
    # comparison to simplification_dict to decide which sites are removed
    if not simplification_dict['C']:
        removal_list += cation_list     
    if not simplification_dict['A']:
        removal_list += anion_list                
    if not simplification_dict['M']:
        removal_list += mobile_list
    if not simplification_dict['N']:
        removal_list += neutral_list
    
    # Special cases for the structures_A and structures_CAN representations
    # Some structures have only Li. For these we are going to handle them as anions (because every representations includes anions)
    if len(structure) == len(mobile_list):
        if not simplification_dict['M']:
            for idx in mobile_list:
                structure.replace(idx, mg.Specie("S", oxidation_state=charge))
    
    # Some structures have only neutrals or cations. For these we are going to handle them as anions (because every representation includes anions)
    elif len(structure) == len(removal_list):
        if len(neutral_list) > 0:
            for idx in neutral_list:
                structure.replace(idx, mg.Specie("S", oxidation_state=charge))
            structure.remove_sites(cation_list+mobile_list)
        elif len(mobile_list) > 0:
            for idx in mobile_list:
                structure.replace(idx, mg.Specie("S", oxidation_state=charge))
            structure.remove_sites(cation_list+neutral_list)
        elif len(cation_list) > 0:
            for idx in cation_list:
                structure.replace(idx, mg.Specie("S", oxidation_state=charge))
            structure.remove_sites(neutral_list+mobile_list)
    
    # otherwise just remove whatever is in the removal list
    else:
        structure.remove_sites(removal_list)
    
    # if simplification_dict indicates that the lattice should be scaled
    if simplification_dict['40']:              
        if scaling_counter > 0:
            structure.scale_lattice(40*scaling_counter)
    
    return structure

# the eight distinct simplification dictionary are hardcoded here
simplification_dict_A = {'C':False, 'A':True, 'M':False, 'N':False, '40':False}
simplification_dict_AM = {'C':False, 'A':True, 'M':True, 'N':False, '40':False}
simplification_dict_CAN = {'C':True, 'A':True, 'M':False, 'N':True, '40':False}
simplification_dict_CAMN = {'C':True, 'A':True, 'M':True, 'N':True, '40':False}
simplification_dict_A40 = {'C':False, 'A':True, 'M':False, 'N':False, '40':True}
simplification_dict_AM40 = {'C':False, 'A':True, 'M':True, 'N':False, '40':True}
simplification_dict_CAN40 = {'C':True, 'A':True, 'M':False, 'N':True, '40':True}
simplification_dict_CAMN40 = {'C':True, 'A':True, 'M':True, 'N':True, '40':True}

# application of the structure_simplification() function, applied once for each simplification dictionary
structures_df['structure_A'] = structures_df['structure'].progress_apply(structure_simplifications, simplification_dict=simplification_dict_A)
structures_df['structure_AM'] = structures_df['structure'].progress_apply(structure_simplifications, simplification_dict=simplification_dict_AM)
structures_df['structure_CAN'] = structures_df['structure'].progress_apply(structure_simplifications, simplification_dict=simplification_dict_CAN)
structures_df['structure_CAMN'] = structures_df['structure'].progress_apply(structure_simplifications, simplification_dict=simplification_dict_CAMN)
structures_df['structure_A40'] = structures_df['structure'].progress_apply(structure_simplifications, simplification_dict=simplification_dict_A40)
structures_df['structure_AM40'] = structures_df['structure'].progress_apply(structure_simplifications, simplification_dict=simplification_dict_AM40)
structures_df['structure_CAN40'] = structures_df['structure'].progress_apply(structure_simplifications, simplification_dict=simplification_dict_CAN40)
structures_df['structure_CAMN40'] = structures_df['structure'].progress_apply(structure_simplifications, simplification_dict=simplification_dict_CAMN40)

# inspection of the dataframe with the eight new simplifications
structures_df.head()

### 1q. Save and reload dataframe

save_path = os.path.join(os.getcwd(), 'semi-supervised_supporting_files/structures_df_3p8_post_sanitize_w_simplifications.pkl')
save_file = open(save_path, 'wb')
pickle.dump(structures_df, save_file)
save_file.close()

save_path = os.path.join(os.getcwd(), 'semi-supervised_supporting_files/structures_df_3p8_post_sanitize_w_simplifications.pkl')
open_file = open(save_path, 'rb')
structures_df = pickle.load(open_file)
open_file.close()

### 1r. Use nglview to manually examine some of the simplifications.

import nglview as nv

structures_df.loc[1000, "structure"]

view = nv.show_pymatgen(structures_df.loc[1000, "structure"])
view._remote_call('setSize', args = ['', '600px'])
view.camera = 'orthographic'
view.add_unitcell()
view

view2 = nv.show_pymatgen(structures_df.loc[1000, "structure_CAMN40"])
view2._remote_call('setSize', args = ['', '600px'])
view2.camera = 'orthographic'
view2.add_unitcell()
view2
